Move getServerSideProps to src/screens
