+++ packages/app-project/src/screens/ProjectHomePage/index.js
@@ -1,1 +1,3 @@
 export { default } from './ProjectHomePageContainer'
+export { getServerSideProps } from './getServerSideProps'
+
